Cleaned up the tabs and files, removed unnecessary files and made things more manageable, We can also fix directories or clean them up
