--- RecipeApp/app/_layout.tsx
@@ -1,13 +1,0 @@
-import { Stack } from 'expo-router';
-
-
-export default function RootLayout() {
-  
-  return (
-    <Stack>
-      <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
-      <Stack.Screen name="+not-found" />
-    </Stack>
-
-  );
-}
